+++ test/IgnoredExactSelectors.spec.ts
@@ -1,0 +1,18 @@
+import * as path from "path";
+
+import { it } from "@jest/globals";
+
+import { assertActualMatchesExpectedAfterPrefixWrap } from "./support/PrefixAssert";
+import { postCSSWithPlugin } from "./support/PluginBootstrap";
+
+const postCSSIgnore = postCSSWithPlugin({
+    ignoredSelectors: ["body"],
+});
+
+it("ignores selectors that are in a ignore list", () => {
+    assertActualMatchesExpectedAfterPrefixWrap(
+        postCSSIgnore,
+        path.join(__dirname, "fixtures", "ignore-exact-selectors-raw.css"),
+        path.join(__dirname, "fixtures", "ignore-exact-selectors-expected.css"),
+    );
+});